pyproject.toml
[project]
name = "magpie"
version = "0.1.0"
description = "LLM-powered options trading system using Alpaca paper trading"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Alpaca
    "alpaca-py>=0.38.0",
    # LLM
    "anthropic>=0.40.0",
    # Database
    "duckdb>=1.2.0",
    # Config & validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    # CLI
    "typer>=0.15.0",
    "rich>=14.0.0",
    # Options math (local Greeks computation)
    "py-vollib>=1.0.1",
    # Utilities
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
    "httpx>=0.28.0",
    "pendulum>=3.0.0",
    "tenacity>=9.0.0",
]

[project.scripts]
magpie = "magpie.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/magpie"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
